Add files via upload
The code is based on Source Code for driving the 28BYJ-48 Stepper Motor https://github.com/BretStateham/28BYJ-48. The modifications are to rotate the sampling tubes, so we are collecting rain water for some minutes on one tube and then rotate to the next one. Once we fill all the tubes, we stop.